requirements.txt
channels==2.1.1
channels-redis==2.2.0
Django==2.0.5
django-cors-headers==2.2.0
django-nose==1.4.5
djangorestframework==3.8.2
django-rq==1.1.0
pytest-asyncio==0.8.0
pytest-django==3.2.1
psycopg2-binary==2.7.4
